// frontend/src/sections/cold-chain-equipment/catalogue.generated.ts
export type AssetPropertiesListResult = {
  assetProperties: ({
  __typename: "AssetPropertyConnector";
} & {
  totalCount: number;
  nodes: Array<{
  id: string;
  key: string;
  name: string;
  valueType: "STRING" | "BOOLEAN" | "INTEGER" | "FLOAT" | "DATE";
  allowedValues: string | null;
  assetClassId: string | null;
  assetCategoryId: string | null;
  assetTypeId: string | null;
}>;
});
};